+++ R/addons-GSVA.R
@@ -7,11 +7,11 @@
 #
 append_transpilers_for_GSVA <- function() {
   ns <- getNamespace("GSVA")
-  ## gsva() in GSVA (< 2.4) relies on BiocParallel::bpiterate()
-  ## which does _not_ support DoparParam by design.
   
   ## These GSVA functions accept BPPARAM via '...' (S4 generics)
   names <- c(
+    ## gsva() in GSVA (< 2.4) relies on BiocParallel::bpiterate()
+    ## which does _not_ support DoparParam by design.
     if (getNamespaceVersion(ns) >= "2.4") "gsva",
     "gsvaRanks", "gsvaScores",
     "spatCor"